Extract App::call tests into a file
