+++ tests/Container/ContainerCallTest.php
@@ -1,0 +1,195 @@
+<?php
+
+namespace Illuminate\Tests\Container;
+
+use Closure;
+use stdClass;
+use PHPUnit\Framework\TestCase;
+use Illuminate\Container\Container;
+
+class ContainerCallTest extends TestCase
+{
+    /**
+     * @expectedException \ReflectionException
+     * @expectedExceptionMessage Function ContainerTestCallStub() does not exist
+     */
+    public function testCallWithAtSignBasedClassReferencesWithoutMethodThrowsException()
+    {
+        $container = new Container;
+        $container->call('ContainerTestCallStub');
+    }
+
+    public function testCallWithAtSignBasedClassReferences()
+    {
+        $container = new Container;
+        $result = $container->call(ContainerTestCallStub::class.'@work', ['foo', 'bar']);
+        $this->assertEquals(['foo', 'bar'], $result);
+
+        $container = new Container;
+        $result = $container->call(ContainerTestCallStub::class.'@inject');
+        $this->assertInstanceOf(ContainerCallConcreteStub::class, $result[0]);
+        $this->assertEquals('taylor', $result[1]);
+
+        $container = new Container;
+        $result = $container->call(ContainerTestCallStub::class.'@inject', ['default' => 'foo']);
+        $this->assertInstanceOf(ContainerCallConcreteStub::class, $result[0]);
+        $this->assertEquals('foo', $result[1]);
+
+        $container = new Container;
+        $result = $container->call(ContainerTestCallStub::class, ['foo', 'bar'], 'work');
+        $this->assertEquals(['foo', 'bar'], $result);
+    }
+
+    public function testCallWithCallableArray()
+    {
+        $container = new Container;
+        $stub = new ContainerTestCallStub;
+        $result = $container->call([$stub, 'work'], ['foo', 'bar']);
+        $this->assertEquals(['foo', 'bar'], $result);
+    }
+
+    public function testCallWithStaticMethodNameString()
+    {
+        $container = new Container;
+        $result = $container->call('Illuminate\Tests\Container\ContainerStaticMethodStub::inject');
+        $this->assertInstanceOf(ContainerCallConcreteStub::class, $result[0]);
+        $this->assertEquals('taylor', $result[1]);
+    }
+
+    public function testCallWithGlobalMethodName()
+    {
+        $container = new Container;
+        $result = $container->call('Illuminate\Tests\Container\containerTestInject');
+        $this->assertInstanceOf(ContainerCallConcreteStub::class, $result[0]);
+        $this->assertEquals('taylor', $result[1]);
+    }
+
+    public function testCallWithBoundMethod()
+    {
+        $container = new Container;
+        $container->bindMethod(ContainerTestCallStub::class.'@unresolvable', function ($stub) {
+            return $stub->unresolvable('foo', 'bar');
+        });
+        $result = $container->call(ContainerTestCallStub::class.'@unresolvable');
+        $this->assertEquals(['foo', 'bar'], $result);
+
+        $container = new Container;
+        $container->bindMethod(ContainerTestCallStub::class.'@unresolvable', function ($stub) {
+            return $stub->unresolvable('foo', 'bar');
+        });
+        $result = $container->call([new ContainerTestCallStub, 'unresolvable']);
+        $this->assertEquals(['foo', 'bar'], $result);
+
+        $container = new Container;
+        $result = $container->call([new ContainerTestCallStub, 'inject'], ['_stub' => 'foo', 'default' => 'bar']);
+        $this->assertInstanceOf(ContainerCallConcreteStub::class, $result[0]);
+        $this->assertEquals('bar', $result[1]);
+
+        $container = new Container;
+        $result = $container->call([new ContainerTestCallStub, 'inject'], ['_stub' => 'foo']);
+        $this->assertInstanceOf(ContainerCallConcreteStub::class, $result[0]);
+        $this->assertEquals('taylor', $result[1]);
+    }
+
+    public function testBindMethodAcceptsAnArray()
+    {
+        $container = new Container;
+        $container->bindMethod([ContainerTestCallStub::class, 'unresolvable'], function ($stub) {
+            return $stub->unresolvable('foo', 'bar');
+        });
+        $result = $container->call(ContainerTestCallStub::class.'@unresolvable');
+        $this->assertEquals(['foo', 'bar'], $result);
+
+        $container = new Container;
+        $container->bindMethod([ContainerTestCallStub::class, 'unresolvable'], function ($stub) {
+            return $stub->unresolvable('foo', 'bar');
+        });
+        $result = $container->call([new ContainerTestCallStub, 'unresolvable']);
+        $this->assertEquals(['foo', 'bar'], $result);
+    }
+
+    public function testClosureCallWithInjectedDependency()
+    {
+        $container = new Container;
+        $container->call(function (ContainerCallConcreteStub $stub) {
+        }, ['foo' => 'bar']);
+
+        $container->call(function (ContainerCallConcreteStub $stub) {
+        }, ['foo' => 'bar', 'stub' => new ContainerCallConcreteStub]);
+    }
+
+    public function testCallWithDependencies()
+    {
+        $container = new Container;
+        $result = $container->call(function (stdClass $foo, $bar = []) {
+            return func_get_args();
+        });
+
+        $this->assertInstanceOf(stdClass::class, $result[0]);
+        $this->assertEquals([], $result[1]);
+
+        $result = $container->call(function (stdClass $foo, $bar = []) {
+            return func_get_args();
+        }, ['bar' => 'taylor']);
+
+        $this->assertInstanceOf(stdClass::class, $result[0]);
+        $this->assertEquals('taylor', $result[1]);
+
+        $stub = new ContainerCallConcreteStub;
+        $result = $container->call(function (stdClass $foo, ContainerCallConcreteStub $bar) {
+            return func_get_args();
+        }, [ContainerCallConcreteStub::class => $stub]);
+
+        $this->assertInstanceOf(stdClass::class, $result[0]);
+        $this->assertSame($stub, $result[1]);
+
+        /*
+         * Wrap a function...
+         */
+        $result = $container->wrap(function (stdClass $foo, $bar = []) {
+            return func_get_args();
+        }, ['bar' => 'taylor']);
+
+        $this->assertInstanceOf(Closure::class, $result);
+        $result = $result();
+
+        $this->assertInstanceOf(stdClass::class, $result[0]);
+        $this->assertEquals('taylor', $result[1]);
+    }
+}
+
+class ContainerTestCallStub
+{
+    public function work()
+    {
+        return func_get_args();
+    }
+
+    public function inject(ContainerCallConcreteStub $stub, $default = 'taylor')
+    {
+        return func_get_args();
+    }
+
+    public function unresolvable($foo, $bar)
+    {
+        return func_get_args();
+    }
+}
+
+class ContainerCallConcreteStub
+{
+    //
+}
+
+function containerTestInject(ContainerCallConcreteStub $stub, $default = 'taylor')
+{
+    return func_get_args();
+}
+
+class ContainerStaticMethodStub
+{
+    public static function inject(ContainerCallConcreteStub $stub, $default = 'taylor')
+    {
+        return func_get_args();
+    }
+}
